Fix issue with Create Recipe and start index
- Corrected the bug that was stopping my create recipe from working ->
the issue was a button that I created in my HTML accidentally had the
same class as another button I had already made. The event listener was
not able to listen to both and therefore crashed.

- Once I finished correcting the create problem, I wrote the code Needed
in my events.js api.js and ui.js for indexing all of a users recipes